+++ tests/regression/query/meta.py
@@ -1,5 +1,5 @@
 '''Test query meta and corner cases'''
-from stdnet import test
+from stdnet import test, QuerySetError
 
 from examples.models import Instrument
 from examples.data import FinanceTest
@@ -18,5 +18,16 @@ class TestMeta(FinanceTest):
         self.assertEqual(list(empty),[])
         self.assertEqual(empty.executed,True)
         self.assertEqual(empty.construct(),empty)
+        
+    def testProperties(self):
+        query = Instrument.objects.query()
+        self.assertFalse(query.executed)
+        
+    def test_getfield(self):
+        query = Instrument.objects.query()
+        self.assertRaises(QuerySetError, query.get_field, 'waaaaaaa')
+        query = query.get_field('id')
+        query2 = query.get_field('id')
+        self.assertEqual(query,query2)